Move the script_runner into its own file
